Provide access to the host jQuery object
This is useful for module loaders that provide separate instances of dependencies to each module, in which case we add to the jQuery object given to us but that is not the jQuery object the user will receive, rendering our plugin registration useless.

Also, in the full build, concat the lib to the end. This is required because the last IIFE is returned, and we want module loaders to return the `datetimepicker` object instead of our last dependency.